Move all widget code into WidgetCommands
